services/ml_service/requirements.txt
# ML Core - Exact versions that work
joblib==1.4.2
catboost==1.2.7
xgboost==2.0.3
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
scipy==1.15.3

# ONNX Runtime - Ultra-fast inference
onnx==1.17.0
onnxruntime-gpu==1.19.2  # Use onnxruntime==1.19.2 for CPU-only
onnxmltools==1.12.0
skl2onnx==1.17.0
tf2onnx==1.16.1  # Optional, for TensorFlow models

# FastAPI Stack
fastapi==0.115.4
uvicorn[standard]==0.32.0
starlette==0.41.3
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.12

# Async HTTP
aiohttp==3.10.10
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
async-timeout==4.0.3
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0
idna==3.11

# Monitoring & Logging
prometheus-client==0.21.0
structlog==24.4.0
rich==13.9.2

# Production Server
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# Utilities
click==8.3.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
exceptiongroup==1.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
Pygments==2.19.2
sniffio==1.3.1
threadpoolctl==3.6.0
typing-extensions==4.15.0

# Data Processing
pyarrow>=14.0,<15.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0

# Optional Data Visualization (for development)
matplotlib==3.10.7
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
pillow==12.0.0
plotly==6.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
pytz==2025.2
tzdata==2025.2

# Scientific computing extras
narwhals==2.10.2
graphviz==0.21
